Refactoring the invite journey to know about usercases and not just cases
