+++ trade_remedies_api/security/services/v2/views.py
@@ -1,0 +1,8 @@
+from config.viewsets import BaseModelViewSet
+from security.models import UserCase
+from security.services.v2.serializers import UserCaseSerializer
+
+
+class UserCaseViewSet(BaseModelViewSet):
+    queryset = UserCase.objects.all()
+    serializer_class = UserCaseSerializer
